feat: 예외 처리 기능 구현

+++ backend/src/main/kotlin/com/woowahan/campus/announcement/feature/DeleteAnnouncement.kt
@@ -1,6 +1,7 @@
 package com.woowahan.campus.announcement.feature
 
 import com.woowahan.campus.announcement.domain.AnnouncementRepository
+import com.woowahan.campus.announcement.exception.AnnouncementNotFoundException
 import openapi.api.DeleteAnnouncementApi
 import org.springframework.http.ResponseEntity
 import org.springframework.transaction.annotation.Transactional
@@ -13,6 +14,10 @@ class DeleteAnnouncement(
 
     @Transactional
     override fun deleteAnnouncement(id: Long, authorization: String): ResponseEntity<Unit> {
+        if (!announcementRepository.existsById(id)) {
+            throw AnnouncementNotFoundException("존재하지 않는 announcement입니다.")
+        }
+
         announcementRepository.deleteById(id)
 
         return ResponseEntity.noContent().build()
